refactor: line up component structure with other components
