+++ packages/select/docs/Select.mdx
@@ -110,6 +110,6 @@ Keep in mind that using disabled in its current form is an anti-pattern. There w
 
 ## Props
 
-```props packages/select/src/Select.tsx
+```props packages/select/src/component.tsx
 
 ```
