+++ packages/select/src/props.tsx
@@ -1,0 +1,54 @@
+import * as React from 'react';
+
+export type SelectProps = {
+    /** Whether the element should receive focus on render. */
+    autoFocus?: boolean;
+
+    /** The `option` elements to populate the select with */
+    children?: React.ReactNode;
+
+    /** Additional CSS class for the container. */
+    className?: string;
+
+    /** The default value (uncontrolled). */
+    defaultValue?: string;
+
+    /** Whether the select is disabled. */
+    disabled?: boolean;
+
+    /**  Renders the field in an invalid state. Often paired together with `helpText` to provide feedback about the error. */
+    invalid?: boolean;
+
+    /** The content to display as the help text. */
+    helpText?: React.ReactNode;
+
+    /** The element's unique identifier. See [MDN](https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/id). */
+    id?: string;
+
+    /** The content to display as the label. */
+    label?: React.ReactNode;
+
+    /** The name of the select element, used when submitting an HTML form. See [MDN](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input#htmlattrdefname). */
+    name?: string;
+
+    /** Handler that is called when the element loses focus. */
+    onBlur?: (e: React.FocusEvent<HTMLSelectElement>) => void;
+
+    /** Handler that is called when the value changes.*/
+    onChange?: (e: React.ChangeEvent<HTMLSelectElement>) => void;
+
+    /** Handler that is called when the element receives focus. */
+    onFocus?: (e: React.FocusEvent<HTMLSelectElement>) => void;
+
+    /** Whether user input is required on the select before form submission. */
+    required?: boolean;
+
+    /** Additional CSS styles for the container. */
+    style?: React.CSSProperties;
+
+    /** The current value (controlled). */
+    value?: string;
+} & Omit<
+    React.PropsWithoutRef<JSX.IntrinsicElements['select']>,
+    'onBlur' | 'onChange' | 'onFocus' | 'value' | 'defaultValue'
+>;
